refactor(backend): move DDD shared kernel from domain/shared to domain/utils
Consolidate directory structure by moving DDD base classes into the
existing domain/utils directory, eliminating the shared/utils split.
