+++ packages/apps/backend/src/domain/utils/aggregate-root.base.ts
