+++ packages/apps/backend/src/domain/utils/entity.base.ts
