+++ packages/apps/backend/src/domain/utils/value-object.base.ts
